+++ src/deluge/modulation/arpeggiator.cpp
@@ -50,6 +50,7 @@ ArpeggiatorSettings::ArpeggiatorSettings() {
 	lockedBassProbabilityValues.fill(0);
 	lockedChordProbabilityValues.fill(0);
 	lockedRatchetProbabilityValues.fill(0);
+	lockedReverseProbabilityValues.fill(0);
 	lockedSpreadVelocityValues.fill(0);
 	lockedSpreadGateValues.fill(0);
 	lockedSpreadOctaveValues.fill(0);
@@ -93,8 +94,6 @@ void ArpeggiatorBase::resetBase() {
 	// Rhythm
 	notesPlayedFromRhythm = 0;
 	lastNormalNotePlayedFromRhythm = 0;
-	// Randomizer
-	notesPlayedFromLockedRandomizer = 0;
 	// Step repeat
 	stepRepeatIndex = 0;
 }
@@ -139,23 +138,29 @@ void ArpeggiatorForDrum::noteOn(ArpeggiatorSettings* settings, int32_t noteCode,
 
 	// Or otherwise, just switch the note on.
 	else {
-		// Apply spread to non-arpeggiated notes
-		int32_t spreadVelocityForCurrentStep = getRandomBipolarProbabilityAmount(settings->spreadVelocity);
+		// Apply randomizer
+		calculateRandomizerAmounts(settings);
+		notesPlayedFromLockedRandomizer = (notesPlayedFromLockedRandomizer + 1) % RANDOMIZER_LOCK_MAX_SAVED_VALUES;
 
-		arpNote.baseVelocity = originalVelocity;
-		// Now apply velocity spread to the base velocity
-		uint8_t velocity = calculateSpreadVelocity(originalVelocity, spreadVelocityForCurrentStep);
-		arpNote.velocity = velocity; // calculated modified velocity
+		if (isPlayNoteForCurrentStep) {
+			// Play a note
 
-		// Set the note to be played
-		noteCodeCurrentlyOnPostArp[0] = noteCode;
-		arpNote.noteCodeOnPostArp[0] = noteCode;
-		for (int32_t n = 1; n < ARP_MAX_INSTRUCTION_NOTES; n++) {
-			// Clean rest of slots
-			noteCodeCurrentlyOnPostArp[n] = ARP_NOTE_NONE;
-			arpNote.noteCodeOnPostArp[n] = ARP_NOTE_NONE;
+			arpNote.baseVelocity = originalVelocity;
+			// Now apply velocity spread to the base velocity
+			uint8_t velocity = calculateSpreadVelocity(originalVelocity, spreadVelocityForCurrentStep);
+			arpNote.velocity = velocity; // calculated modified velocity
+
+			// Set the note to be played
+			noteCodeCurrentlyOnPostArp[0] = noteCode;
+			arpNote.noteCodeOnPostArp[0] = noteCode;
+			for (int32_t n = 1; n < ARP_MAX_INSTRUCTION_NOTES; n++) {
+				// Clean rest of slots
+				noteCodeCurrentlyOnPostArp[n] = ARP_NOTE_NONE;
+				arpNote.noteCodeOnPostArp[n] = ARP_NOTE_NONE;
+			}
+			instruction->invertReverse = isPlayReverseForCurrentStep;
+			instruction->arpNoteOn = &arpNote;
 		}
-		instruction->arpNoteOn = &arpNote;
 	}
 }
 
@@ -286,23 +291,29 @@ noteInserted:
 		// Don't do the note-on now, it'll happen automatically at next render
 	}
 	else {
-		// Apply spread to non-arpeggiated notes
-		int32_t spreadVelocityForCurrentStep = getRandomBipolarProbabilityAmount(settings->spreadVelocity);
+		// Apply randomizer
+		calculateRandomizerAmounts(settings);
+		notesPlayedFromLockedRandomizer = (notesPlayedFromLockedRandomizer + 1) % RANDOMIZER_LOCK_MAX_SAVED_VALUES;
 
-		arpNote->baseVelocity = originalVelocity;
-		// Now apply velocity spread to the base velocity
-		uint8_t velocity = calculateSpreadVelocity(originalVelocity, spreadVelocityForCurrentStep);
-		arpNote->velocity = velocity; // calculated modified velocity
+		if (isPlayNoteForCurrentStep) {
+			// Play a note
 
-		// Set the note to be played
-		noteCodeCurrentlyOnPostArp[0] = noteCode;
-		arpNote->noteCodeOnPostArp[0] = noteCode;
-		for (int32_t n = 1; n < ARP_MAX_INSTRUCTION_NOTES; n++) {
-			// Clean rest of chord note slots
-			noteCodeCurrentlyOnPostArp[n] = ARP_NOTE_NONE;
-			arpNote->noteCodeOnPostArp[n] = ARP_NOTE_NONE;
+			arpNote->baseVelocity = originalVelocity;
+			// Now apply velocity spread to the base velocity
+			uint8_t velocity = calculateSpreadVelocity(originalVelocity, spreadVelocityForCurrentStep);
+			arpNote->velocity = velocity; // calculated modified velocity
+
+			// Set the note to be played
+			noteCodeCurrentlyOnPostArp[0] = noteCode;
+			arpNote->noteCodeOnPostArp[0] = noteCode;
+			for (int32_t n = 1; n < ARP_MAX_INSTRUCTION_NOTES; n++) {
+				// Clean rest of chord note slots
+				noteCodeCurrentlyOnPostArp[n] = ARP_NOTE_NONE;
+				arpNote->noteCodeOnPostArp[n] = ARP_NOTE_NONE;
+			}
+			instruction->invertReverse = isPlayReverseForCurrentStep;
+			instruction->arpNoteOn = arpNote;
 		}
-		instruction->arpNoteOn = arpNote;
 	}
 }
 
@@ -477,6 +488,12 @@ bool ArpeggiatorBase::evaluateBassProbability(bool isRatchet) {
 	return isRatchet ? lastNormalNotePlayedFromBassProbability : isPlayBassForCurrentStep;
 }
 
+// Returns if the arpeggiator should play the bass note instead of the normal note
+bool ArpeggiatorBase::evaluateReverseProbability(bool isRatchet) {
+	// If it is a rachet, use the last value, but it it is not a ratchet, use the calculated value
+	return isRatchet ? lastNormalNotePlayedFromReverseProbability : isPlayReverseForCurrentStep;
+}
+
 // Returns if the arpeggiator should play a chord instead of the normal note
 bool ArpeggiatorBase::evaluateChordProbability(bool isRatchet) {
 	// If it is a rachet, use the last value, but it it is not a ratchet, use the calculated value
@@ -486,7 +503,8 @@ bool ArpeggiatorBase::evaluateChordProbability(bool isRatchet) {
 // Returns if note should be played
 void ArpeggiatorBase::executeArpStep(ArpeggiatorSettings* settings, uint8_t numActiveNotes, bool isRatchet,
                                      uint32_t maxSequenceLength, uint32_t rhythm, bool shouldCarryOnRhythmNote,
-                                     bool shouldPlayNote, bool shouldPlayBassNote, bool shouldPlayChordNote) {
+                                     bool shouldPlayNote, bool shouldPlayBassNote, bool shouldPlayReverseNote,
+                                     bool shouldPlayChordNote) {
 
 	// Here we reset the arpeggiator sequence based on several possible conditions
 	if (settings->flagForceArpRestart) {
@@ -529,10 +547,11 @@ void ArpeggiatorBase::executeArpStep(ArpeggiatorSettings* settings, uint8_t numA
 			else {
 				// Move indexes to the next note in the sequence, according to OctaveMode, NoteMode and StepRepeat
 				calculateNextNoteAndOrOctave(settings, numActiveNotes);
-
-				// Increase pattern related indexes
-				increasePatternIndexes(settings->numStepRepeats);
 			}
+
+			// Increase pattern related indexes
+			increasePatternIndexes(settings->numStepRepeats);
+
 			// As we have set a note and an octave, we can say we have played a note
 			playedFirstArpeggiatedNoteYet = true;
 
@@ -544,6 +563,9 @@ void ArpeggiatorBase::executeArpStep(ArpeggiatorSettings* settings, uint8_t numA
 
 			// Save last note played from probability
 			lastNormalNotePlayedFromBassProbability = shouldPlayBassNote;
+
+			// Save last note played from probability
+			lastNormalNotePlayedFromReverseProbability = shouldPlayReverseNote;
 
 			// Save last note played from probability
 			lastNormalNotePlayedFromChordProbability = shouldPlayChordNote;
@@ -593,11 +615,11 @@ void ArpeggiatorForDrum::switchNoteOn(ArpeggiatorSettings* settings, ArpReturnIn
 	}
 	bool shouldPlayNote = evaluateNoteProbability(isRatchet);
 	bool shouldPlayBassNote = evaluateBassProbability(isRatchet);
-	bool shouldPlayChordNote = evaluateChordProbability(isRatchet);
+	bool shouldPlayReverseNote = evaluateReverseProbability(isRatchet);
 
 	// Execute all the step calculations
 	executeArpStep(settings, chordTypeNoteCount[settings->chordTypeIndex], isRatchet, maxSequenceLength, rhythm,
-	               shouldCarryOnRhythmNote, shouldPlayNote, shouldPlayBassNote, shouldPlayChordNote);
+	               shouldCarryOnRhythmNote, shouldPlayNote, shouldPlayBassNote, shouldPlayReverseNote, false);
 
 	if (shouldCarryOnRhythmNote && shouldPlayNote) {
 		// Set Gate as active
@@ -663,13 +685,17 @@ void ArpeggiatorForDrum::switchNoteOn(ArpeggiatorSettings* settings, ArpReturnIn
 			noteCodeCurrentlyOnPostArp[n] = ARP_NOTE_NONE;
 			arpNote.noteCodeOnPostArp[n] = ARP_NOTE_NONE;
 		}
+		instruction->invertReverse = shouldPlayReverseNote;
 		instruction->arpNoteOn = &arpNote;
 	}
 }
 
 bool ArpeggiatorBase::getRandomProbabilityResult(uint32_t value) {
 	if (value == 0) {
-		return false;
+		return false; // If value at min probability, return always false
+	}
+	if (value == 4294967295u) {
+		return true; // If value at max probability, return always true
 	}
 	uint32_t randomChance = random(65535);
 	return (value >> 16) >= randomChance;
@@ -711,6 +737,13 @@ void ArpeggiatorBase::calculateRandomizerAmounts(ArpeggiatorSettings* settings) 
 				settings->lockedBassProbabilityValues[i] = getRandomProbabilityResult(settings->bassProbability);
 			}
 			settings->lastLockedBassProbabilityParameterValue = settings->bassProbability;
+		}
+		if (resetLockedRandomizerValuesNextTime
+		    || settings->lastLockedReverseProbabilityParameterValue != settings->reverseProbability) {
+			for (int i = 0; i < RANDOMIZER_LOCK_MAX_SAVED_VALUES; i++) {
+				settings->lockedReverseProbabilityValues[i] = getRandomProbabilityResult(settings->reverseProbability);
+			}
+			settings->lastLockedReverseProbabilityParameterValue = settings->reverseProbability;
 		}
 		if (resetLockedRandomizerValuesNextTime
 		    || settings->lastLockedChordProbabilityParameterValue != settings->chordProbability) {
@@ -756,6 +789,9 @@ void ArpeggiatorBase::calculateRandomizerAmounts(ArpeggiatorSettings* settings) 
 		isPlayBassForCurrentStep =
 		    settings->lockedBassProbabilityValues[notesPlayedFromLockedRandomizer % RANDOMIZER_LOCK_MAX_SAVED_VALUES]
 		    != 0;
+		isPlayReverseForCurrentStep =
+		    settings->lockedReverseProbabilityValues[notesPlayedFromLockedRandomizer % RANDOMIZER_LOCK_MAX_SAVED_VALUES]
+		    != 0;
 		isPlayChordForCurrentStep =
 		    settings->lockedChordProbabilityValues[notesPlayedFromLockedRandomizer % RANDOMIZER_LOCK_MAX_SAVED_VALUES]
 		    != 0;
@@ -773,6 +809,7 @@ void ArpeggiatorBase::calculateRandomizerAmounts(ArpeggiatorSettings* settings) 
 		// Lively create new randomized values on the fly each time a note is played
 		isPlayNoteForCurrentStep = getRandomProbabilityResult(settings->noteProbability);
 		isPlayBassForCurrentStep = getRandomProbabilityResult(settings->bassProbability);
+		isPlayReverseForCurrentStep = getRandomProbabilityResult(settings->reverseProbability);
 		isPlayChordForCurrentStep = getRandomProbabilityResult(settings->chordProbability);
 		isPlayRatchetForCurrentStep = getRandomProbabilityResult(settings->ratchetProbability);
 		spreadVelocityForCurrentStep = getRandomBipolarProbabilityAmount(settings->spreadVelocity);
@@ -1032,11 +1069,13 @@ void Arpeggiator::switchNoteOn(ArpeggiatorSettings* settings, ArpReturnInstructi
 	}
 	bool shouldPlayNote = evaluateNoteProbability(isRatchet);
 	bool shouldPlayBassNote = evaluateBassProbability(isRatchet);
+	bool shouldPlayReverseNote = evaluateReverseProbability(isRatchet);
 	bool shouldPlayChordNote = evaluateChordProbability(isRatchet);
 
 	// Execute all the step calculations
 	executeArpStep(settings, (uint8_t)notes.getNumElements(), isRatchet, maxSequenceLength, rhythm,
-	               shouldCarryOnRhythmNote, shouldPlayNote, shouldPlayBassNote, shouldPlayChordNote);
+	               shouldCarryOnRhythmNote, shouldPlayNote, shouldPlayBassNote, shouldPlayReverseNote,
+	               shouldPlayChordNote);
 
 	// Clamp the index to real range
 	whichNoteCurrentlyOnPostArp = std::clamp<int16_t>(whichNoteCurrentlyOnPostArp, 0, notes.getNumElements() - 1);
@@ -1180,6 +1219,7 @@ void Arpeggiator::switchNoteOn(ArpeggiatorSettings* settings, ArpReturnInstructi
 		for (int32_t n = 0; n < ARP_MAX_INSTRUCTION_NOTES; n++) {
 			arpNote->noteCodeOnPostArp[n] = noteCodeCurrentlyOnPostArp[n];
 		}
+		instruction->invertReverse = shouldPlayReverseNote;
 		instruction->arpNoteOn = arpNote;
 	}
 }
@@ -1262,6 +1302,9 @@ void ArpeggiatorBase::render(ArpeggiatorSettings* settings, ArpReturnInstruction
 // May switch notes on and/or off.
 int32_t ArpeggiatorBase::doTickForward(ArpeggiatorSettings* settings, ArpReturnInstruction* instruction,
                                        uint32_t clipCurrentPos, bool currentlyPlayingReversed) {
+	if (clipCurrentPos == 0) {
+		notesPlayedFromLockedRandomizer = 0;
+	}
 	// Make sure we actually intended to sync
 	if (settings->mode == ArpMode::OFF || (settings->syncLevel == 0u)) {
 		return 2147483647;
@@ -1341,6 +1384,7 @@ void ArpeggiatorSettings::cloneFrom(ArpeggiatorSettings const* other) {
 	ratchetAmount = other->ratchetAmount;
 	noteProbability = other->noteProbability;
 	bassProbability = other->bassProbability;
+	reverseProbability = other->reverseProbability;
 	chordProbability = other->chordProbability;
 	ratchetProbability = other->ratchetProbability;
 	spreadVelocity = other->spreadVelocity;
@@ -1372,6 +1416,13 @@ bool ArpeggiatorSettings::readCommonTagsFromFile(Deserializer& reader, char cons
 	else if (!strcmp(tagName, "lockedBassProbArray")) {
 		int len = reader.readTagOrAttributeValueHexBytes((uint8_t*)lockedBassProbabilityValues.data(),
 		                                                 lockedBassProbabilityValues.size());
+	}
+	else if (!strcmp(tagName, "lastLockedReverseProb")) {
+		lastLockedReverseProbabilityParameterValue = reader.readTagOrAttributeValueInt();
+	}
+	else if (!strcmp(tagName, "lockeReverseProbArray")) {
+		int len = reader.readTagOrAttributeValueHexBytes((uint8_t*)lockedReverseProbabilityValues.data(),
+		                                                 lockedReverseProbabilityValues.size());
 	}
 	else if (!strcmp(tagName, "lastLockedChordProb")) {
 		lastLockedChordProbabilityParameterValue = reader.readTagOrAttributeValueInt();
@@ -1479,6 +1530,9 @@ bool ArpeggiatorSettings::readNonAudioTagsFromFile(Deserializer& reader, char co
 	else if (!strcmp(tagName, "bassProbability")) {
 		bassProbability = reader.readTagOrAttributeValueInt();
 	}
+	else if (!strcmp(tagName, "reverseProbability")) {
+		reverseProbability = reader.readTagOrAttributeValueInt();
+	}
 	else if (!strcmp(tagName, "chordProbability")) {
 		chordProbability = reader.readTagOrAttributeValueInt();
 	}
@@ -1546,6 +1600,10 @@ void ArpeggiatorSettings::writeCommonParamsToFile(Serializer& writer, Song* song
 	writer.writeAttribute("lastLockedBassProb", lastLockedBassProbabilityParameterValue);
 	writer.writeAttributeHexBytes("lockedBassProbArray", (uint8_t*)lockedBassProbabilityValues.data(),
 	                              lockedBassProbabilityValues.size());
+	// Reverse probability
+	writer.writeAttribute("lastLockedReverseProb", lastLockedReverseProbabilityParameterValue);
+	writer.writeAttributeHexBytes("lockedReverseProbArray", (uint8_t*)lockedReverseProbabilityValues.data(),
+	                              lockedReverseProbabilityValues.size());
 	// Chord probability
 	writer.writeAttribute("lastLockedChordProb", lastLockedChordProbabilityParameterValue);
 	writer.writeAttributeHexBytes("lockedChordProbArray", (uint8_t*)lockedChordProbabilityValues.data(),
@@ -1577,6 +1635,7 @@ void ArpeggiatorSettings::writeNonAudioParamsToFile(Serializer& writer) {
 	// tag)
 	writer.writeAttribute("noteProbability", noteProbability);
 	writer.writeAttribute("bassProbability", bassProbability);
+	writer.writeAttribute("reverseProbability", reverseProbability);
 	writer.writeAttribute("chordProbability", chordProbability);
 	writer.writeAttribute("ratchetProbability", ratchetProbability);
 	writer.writeAttribute("ratchetAmount", ratchetAmount);