+++ src/deluge/modulation/arpeggiator.h
@@ -88,6 +88,7 @@ public:
 	// Spread last lock
 	uint32_t lastLockedNoteProbabilityParameterValue{0};
 	uint32_t lastLockedBassProbabilityParameterValue{0};
+	uint32_t lastLockedReverseProbabilityParameterValue{0};
 	uint32_t lastLockedChordProbabilityParameterValue{0};
 	uint32_t lastLockedRatchetProbabilityParameterValue{0};
 	uint32_t lastLockedSpreadVelocityParameterValue{0};
@@ -97,6 +98,7 @@ public:
 	// Pre-calculated randomized values for each parameter
 	std::array<int8_t, RANDOMIZER_LOCK_MAX_SAVED_VALUES> lockedNoteProbabilityValues;
 	std::array<int8_t, RANDOMIZER_LOCK_MAX_SAVED_VALUES> lockedBassProbabilityValues;
+	std::array<int8_t, RANDOMIZER_LOCK_MAX_SAVED_VALUES> lockedReverseProbabilityValues;
 	std::array<int8_t, RANDOMIZER_LOCK_MAX_SAVED_VALUES> lockedChordProbabilityValues;
 	std::array<int8_t, RANDOMIZER_LOCK_MAX_SAVED_VALUES> lockedRatchetProbabilityValues;
 	std::array<int8_t, RANDOMIZER_LOCK_MAX_SAVED_VALUES> lockedSpreadVelocityValues;
@@ -118,6 +120,7 @@ public:
 	uint32_t ratchetAmount{0};
 	uint32_t noteProbability{4294967295u}; // Default to 25 if not set in XML
 	uint32_t bassProbability{0};
+	uint32_t reverseProbability{0};
 	uint32_t chordProbability{0};
 	uint32_t ratchetProbability{0};
 	uint32_t spreadVelocity{0};
@@ -149,6 +152,7 @@ class ArpReturnInstruction {
 public:
 	ArpReturnInstruction() {
 		sampleSyncLengthOn = 0;
+		invertReverse = false;
 		arpNoteOn = nullptr;
 		outputMIDIChannelOff.fill(MIDI_CHANNEL_NONE);
 		noteCodeOffPostArp.fill(ARP_NOTE_NONE);
@@ -158,6 +162,11 @@ public:
 	// notes are still playing (e.g. for mono note priority)
 	uint32_t sampleSyncLengthOn; // This defaults to zero, or may be overwritten by the caller to the Arp - and then
 	                             // the Arp itself may override that.
+
+	// If set to true by the arpeggiator, the reverse SAMPLE flag will be inverted for the next voices to be played
+	// and restored back to normal afterwards
+	bool invertReverse;
+
 	ArpNote* arpNoteOn;
 
 	// And these are only valid if doing a note-off
@@ -210,10 +219,13 @@ public:
 	bool lastNormalNotePlayedFromNoteProbability = true;
 
 	// Bass probability state
-	bool lastNormalNotePlayedFromBassProbability = true;
+	bool lastNormalNotePlayedFromBassProbability = false;
+
+	// Reverse probability state
+	bool lastNormalNotePlayedFromReverseProbability = false;
 
 	// Chord probability state
-	bool lastNormalNotePlayedFromChordProbability = true;
+	bool lastNormalNotePlayedFromChordProbability = false;
 
 	// Step repeat state
 	int32_t stepRepeatIndex = 0;
@@ -230,6 +242,7 @@ public:
 	// Calculated spread amounts
 	bool isPlayNoteForCurrentStep = true;
 	bool isPlayBassForCurrentStep = false;
+	bool isPlayReverseForCurrentStep = false;
 	bool isPlayChordForCurrentStep = false;
 	bool isPlayRatchetForCurrentStep = false;
 	int32_t spreadVelocityForCurrentStep = 0;
@@ -244,13 +257,14 @@ protected:
 	void resetRatchet();
 	void executeArpStep(ArpeggiatorSettings* settings, uint8_t numActiveNotes, bool isRatchet,
 	                    uint32_t maxSequenceLength, uint32_t rhythm, bool shouldCarryOnRhythmNote, bool shouldPlayNote,
-	                    bool shouldPlayBassNote, bool shouldPlayChordNote);
+	                    bool shouldPlayBassNote, bool shouldPlayReverseNote, bool shouldPlayChordNote);
 	void increasePatternIndexes(uint8_t numStepRepeats);
 	void increaseSequenceIndexes(uint32_t maxSequenceLength, uint32_t rhythm);
 	void maybeSetupNewRatchet(ArpeggiatorSettings* settings);
 	bool evaluateRhythm(uint32_t rhythm, bool isRatchet);
 	bool evaluateNoteProbability(bool isRatchet);
 	bool evaluateBassProbability(bool isRatchet);
+	bool evaluateReverseProbability(bool isRatchet);
 	bool evaluateChordProbability(bool isRatchet);
 	uint32_t calculateSpreadVelocity(uint8_t velocity, int32_t spreadVelocityForCurrentStep);
 	int32_t getOctaveDirection(ArpeggiatorSettings* settings);